refactor(teardown): fix imports
