+++ tests/dashboard/teams/teams-request-access.spec.js
@@ -19,7 +19,7 @@ const {
   checkYourPenpotViewModeConfirmAccessText,
   checkSigningText,
 } = require('../../../helpers/gmail.js');
-const { createTeamName } = require('../../../helpers/teams/create-team-name.js');
+const { createTeamName } = require('../../../helpers/teams/create-team-name');
 
 // Set up Team and File
 async function setupTeamAndFile(page, teamName) {
